+++ agroconnect/lib/main.dart
@@ -1,4 +1,4 @@
-import 'package:agroconnect/main_navigation.dart';
+import 'package:agroconnect/pages/main_navigation.dart';
 import 'package:flutter/material.dart';
 import 'package:google_fonts/google_fonts.dart';
 import 'package:firebase_core/firebase_core.dart';
