+++ agroconnect/lib/services/dummy_client_data.dart
@@ -28,7 +28,7 @@ class DummyClientData {
         phoneNumber: '+351912345678',
         updatedAt: null,
         name: 'João Silva',
-        imagePath: 'assets/images/client1.jpg',
+        imagePath: 'assets/user-images/client1.jpg',
         email: 'joao@email.com',
         city: 'Lisboa',
         parish: 'Alvalade',
@@ -42,7 +42,7 @@ class DummyClientData {
         phoneNumber: '+351923456789',
         updatedAt: null,
         name: 'Maria Santos',
-        imagePath: 'assets/images/client2.jpg',
+        imagePath: 'assets/user-images/client2.jpg',
         email: 'maria@email.com',
         city: 'Porto',
         parish: 'Cedofeita',
@@ -56,7 +56,7 @@ class DummyClientData {
       phoneNumber: '+351912345678',
       updatedAt: null,
       name: 'João Ferreira',
-      imagePath: 'assets/images/client3.jpg',
+      imagePath: 'assets/user-images/client3.jpg',
       email: 'joao@email.com',
       city: 'Lisboa',
       parish: 'Alvalade',
@@ -70,7 +70,7 @@ class DummyClientData {
       phoneNumber: '+351934567890',
       updatedAt: null,
       name: 'Ana Oliveira',
-      imagePath: 'assets/images/client4.jpg',
+      imagePath: 'assets/user-images/client4.jpg',
       email: 'ana@email.com',
       city: 'Coimbra',
       parish: 'Sé Nova',
@@ -84,7 +84,7 @@ class DummyClientData {
       phoneNumber: '+351926789012',
       updatedAt: null,
       name: 'Carlos Silva',
-      imagePath: 'assets/images/client5.jpg',
+      imagePath: 'assets/user-images/client5.jpg',
       email: 'carlos@email.com',
       city: 'Braga',
       parish: 'São Vicente',
@@ -98,7 +98,7 @@ class DummyClientData {
       phoneNumber: '+351938765432',
       updatedAt: null,
       name: 'Sofia Costa',
-      imagePath: 'assets/images/client6.jpg',
+      imagePath: 'assets/user-images/client6.jpg',
       email: 'sofia@email.com',
       city: 'Funchal',
       parish: 'Santa Luzia',
@@ -112,7 +112,7 @@ class DummyClientData {
       phoneNumber: '+351911223344',
       updatedAt: null,
       name: 'Ricardo Mendes',
-      imagePath: 'assets/images/client7.jpg',
+      imagePath: 'assets/user-images/client7.jpg',
       email: 'ricardo@email.com',
       city: 'Aveiro',
       parish: 'Glória e Vera Cruz',
@@ -126,7 +126,7 @@ class DummyClientData {
       phoneNumber: '+351924567890',
       updatedAt: null,
       name: 'Beatriz Almeida',
-      imagePath: 'assets/images/client8.jpg',
+      imagePath: 'assets/user-images/client8.jpg',
       email: 'beatriz@email.com',
       city: 'Setúbal',
       parish: 'São Sebastião',
@@ -140,7 +140,7 @@ class DummyClientData {
       phoneNumber: '+351917654321',
       updatedAt: null,
       name: 'Tiago Rocha',
-      imagePath: 'assets/images/client9.jpg',
+      imagePath: 'assets/user-images/client9.jpg',
       email: 'tiago@email.com',
       city: 'Leiria',
       parish: 'Marrazes',
@@ -154,7 +154,7 @@ class DummyClientData {
       phoneNumber: '+351935678901',
       updatedAt: null,
       name: 'Helena Pinto',
-      imagePath: 'assets/images/client10.jpg',
+      imagePath: 'assets/user-images/client10.jpg',
       email: 'helena@email.com',
       city: 'Viseu',
       parish: 'Coração de Jesus',
